% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cleanQMJMonthly.R
\name{cleanQMJMonthly}
\alias{cleanQMJMonthly}
\title{Reads in, cleans, and subdivides monthly QMJ data set in data folder.}
\usage{
cleanQMJMonthly()
}
\description{
Reads in, cleans, and subdivides monthly QMJ data set in data folder.
}

