\name{mcmc1dRDE-package}
\alias{mcmc1dRDE-package}
\alias{mcmc1dRDE}
\docType{package}
\title{
\packageTitle{mcmc1dRDE}
}
\description{
\packageDescription{mcmc1dRDE}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{mcmc1dRDE}

Maintainer: \packageMaintainer{mcmc1dRDE}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = mcmc1dRDE)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{mcmc1dRDE}}
%\section{Documentation Index}{\packageIndices{mcmc1dRDE}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
