% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/updateData.R
\name{updateData}
\alias{updateData}
\title{Updates all raw data sets in data and calls cleanAll() if clean is true.}
\usage{
updateData(clean = TRUE)
}
\arguments{
\item{clean}{Cleans updated data by default.}
}
\description{
Updates all raw data sets in data and calls cleanAll() if clean is true.
}

